cardTableBarrierSet.cpp

+++ src/hotspot/share/c1/c1_Instruction.hpp
@@ -1718,7 +1718,7 @@ LEAF(BlockBegin, StateSplit)
   void set_first_lir_instruction_id(int id)      { _first_lir_instruction_id = id;  }
   void set_last_lir_instruction_id(int id)       { _last_lir_instruction_id = id;  }
   void increment_total_preds(int n = 1)          { _total_preds += n; }
-  void init_stores_to_locals(int locals_count)   { _stores_to_locals.initialize(locals_count); }
+  void init_stores_to_locals(int locals_count)   { _stores_to_locals.initialize(static_cast<uint>(locals_count)); }
 
   // generic
   virtual void state_values_do(ValueVisitor* f);
